Relabel the internal list

The entries of the internal list are either digits (0-x) or expressions. To ensure a valid mapping of these values to an input vector (as done when an optimizer provides input parameters), we have to linearize the list (done whenever modifying the list, see the different flex_* functions )

Usage

linearize_internal_list(internal_list)

Arguments

internal_list

the internal list, with entries for each parameter x condition combination

Value

another list, but with remapped digits in increasing order, while leaving expressions or digits of 0 untouched.

See also