Access the Probability Density Functions of a Model

Functions to obtain the probability density functions (PDFs) of a model. These PDFs represent the first-passage-time.

Usage

pdfs(object, ...)

# S3 method for class 'drift_dm'
pdfs(object, ...)

Arguments

object

an object of type drift_dm.

...

additional arguments passed down to the specific method.

Value

A list with named elements corresponding to the conditions of a model (see conds()). Each of these elements is another list, containing the entries pdf_u and pdf_l, which are numeric vectors for the PDFs of the upper and lower boundary, respectively.

Details

If the model has not been evaluated, re_evaluate_model() is called before returning the PDFs.

See also

Examples

# get a pre-built model for demonstration purpose
a_model <- dmc_dm(dx = .0025, dt = .0025)
str(pdfs(a_model))
#> List of 2
#>  $ comp  :List of 2
#>   ..$ pdf_u: num [1:1201] 1e-10 1e-10 1e-10 1e-10 1e-10 ...
#>   ..$ pdf_l: num [1:1201] 1e-10 1e-10 1e-10 1e-10 1e-10 ...
#>  $ incomp:List of 2
#>   ..$ pdf_u: num [1:1201] 1e-10 1e-10 1e-10 1e-10 1e-10 ...
#>   ..$ pdf_l: num [1:1201] 1e-10 1e-10 1e-10 1e-10 1e-10 ...