Create the Shrinking Spotlight Model

This function creates a drift_dm object that corresponds to a simple version of the shrinking spotlight model by Whiteetal.2011;textualdRiftDM.

Usage

ssp_dm(
  instr = NULL,
  obs_data = NULL,
  sigma = 1,
  t_max = 3,
  dt = 0.001,
  dx = 0.001,
  b_coding = NULL
)

Arguments

instr

optional string with additional "instructions", see modify_flex_prms() and the Details below.

obs_data

data.frame, an optional data.frame with the observed data. See obs_data.

sigma, t_max, dt, dx

numeric, providing the settings for the diffusion constant and discretization (see drift_dm)

b_coding

list, an optional list with the boundary encoding (see b_coding)

Value

An object of type drift_dm (parent class) and ssp_dm (child class), created by the function drift_dm().

Details

The shrinking spotlight model is a model developed for the flanker task.

It has the following properties (see component_shelf):

a constant boundary (parameter b)

a constant starting point in between the decision boundaries

an evidence accumulation process that is driven by an attentional spotlight that covers both the flankers and the target. The area that covers the flankers and target is modeled by normal distribution with mean 0:

At the beginning of the trial attention is wide-spread, and the width at t=0 is the standard deviation sd_0

As the trial progresses in time, the attentional spotlight narrows, reflected by a linear decline of the standard deviation with rate r (to a minimum of 0.001).

the attention attributed to both the flankers and the target is scaled by p which controls the strength of evidence accumulation

A non-decision time that follows a truncated normal distribution with mean non_dec and standard deviation sd_non_dec.

The model also contains the auxiliary parameter sign, which is used to control the influence of the flankers across conditions. It is not really a parameter and should not be estimated!

Per default, the parameter r is assumed to be fixed (i.e., is not estimated freely). The model also contains the custom parameter interf_t, quantifying the interference time (sd_0 / r).

References

Whiteetal.2011dRiftDM

Examples

# the model with default settings
my_model <- ssp_dm()

# the model with a more coarse discretization
my_model <- ssp_dm(
  t_max = 1.5,
  dx = .0025,
  dt = .0025
)