Summary and Printing for fits_ids_dm Objects

Methods for summarizing and printing objects of the class fits_ids_dm, which contain multiple fits across individuals.

Usage

# S3 method for class 'summary.fits_ids_dm'
print(x, ..., just_header = FALSE, round_digits = drift_dm_default_rounding())

# S3 method for class 'fits_ids_dm'
summary(object, ..., select_unique = FALSE)

Arguments

x

an object of class summary.fits_ids_dm.

...

additional arguments (currently unused).

just_header

logical, if TRUE only print the header information without details. Default is FALSE.

round_digits

an integer, specifying the number of decimal places for rounding in the printed summary. Default is 3.

object

an object of class fits_ids_dm, generated by a call to load_fits_ids.

select_unique

logical, passed to coef.drift_dm().

Value

summary.fits_ids_dm() returns a list of class summary.fits_ids_dm (see the Details section summarizing each entry of this list).

print.summary.fits_ids_dm() returns invisibly the summary.fits_ids_dm object.

Details

The summary.fits_ids_dm function creates a summary object. The contents of this summary object depends on whether the user supplies a fits_ids_dm object that was created with estimate_dm() or the deprecated function load_fits_ids().

In the first case, the object contains:

summary_drift_dm_obj: A list with information about the underlying drift diffusion model (as returned by summary.drift_dm()).

prms: All parameter values across all conditions (essentially a call to coef() with the argument select_unique = FALSE).

stats: A named list of matrices for each condition, including mean and standard error for each parameter.

obs_data: A list providing the number of individual participants and the average number of trials per condition across participants.

optimizer: A string of the optimizer that was used

conv_info: A list providing a summary of the convergance and messages for all IDs

In the second case, the object contains:

lower and upper: Lower and upper bounds of the search space.

model_type: Description of the model type, based on class information.

prms: All parameter values across all conditions (essentially a call to coef() with the argument select_unique = FALSE).

stats: A named list of matrices for each condition, including mean and standard error for each parameter.

N: The number of individuals.

The print.summary.fits_ids_dm function displays the summary object in a formatted manner.

Examples

# get an auxiliary object of type fits_ids_dm for demonstration purpose
all_fits <- get_example_fits("fits_ids_dm")
sum_obj <- summary(all_fits)
print(sum_obj, round_digits = 2)
#> Fit approach: separately - classical
#> Fitted model type: dmc_dm, drift_dm
#> Optimizer: nmkb 
#> Convergence: TRUE 
#> N Individuals: 3 
#> Average Trial Numbers:
#>  168 trials comp; 168 trials incomp
#> Cost Function: neg_log_like
#> 
#> Parameter Summary: comp 
#>          muc    b non_dec sd_non_dec  tau a    A alpha peak_l
#> mean    4.79 0.47    0.32       0.03 0.07 2 0.12  6.32   0.07
#> std_err 0.44 0.06    0.01       0.01 0.02 0 0.03  1.24   0.02
#> 
#> Parameter Summary: incomp 
#>          muc    b non_dec sd_non_dec  tau a     A alpha peak_l
#> mean    4.79 0.47    0.32       0.03 0.07 2 -0.12  6.32   0.07
#> std_err 0.44 0.06    0.01       0.01 0.02 0  0.03  1.24   0.02
#> 
#> -------
#> Deriving PDFS:
#>   solver: kfe
#>   values: sigma=1, t_max=3, dt=0.0075, dx=0.02, nt=400, nx=100