Summary and Printing for fits_ids_dm Objects

Methods for summarizing and printing objects of the class fits_ids_dm, which contain multiple fits across individuals.

Usage

# S3 method for class 'summary.fits_ids_dm'
print(x, ..., round_digits = drift_dm_default_rounding())

# S3 method for class 'fits_ids_dm'
summary(object, ...)

Arguments

x

an object of class summary.fits_ids_dm.

...

additional arguments

round_digits

integer, specifying the number of decimal places for rounding in the printed summary. Default is set to 3.

object

an object of class fits_ids_dm, generated by a call to load_fits_ids.

Value

summary.fits_ids_dm() returns a list of class summary.fits_ids_dm (see the Details section summarizing each entry of this list).

print.summary.fits_ids_dm() returns invisibly the summary.fits_ids_dm object.

Details

The summary.fits_ids_dm function creates a summary object containing:

fit_procedure_name: The name of the fit procedure used.

time_call: Timestamp of the last fit procedure call.

lower and upper: Lower and upper bounds of the search space.

model_type: Description of the model type, based on class information.

prms: All parameter values across all conditions (essentially a call to coef() with the argument select_unique = FALSE).

stats: A named list of matrices for each condition, including mean and standard error for each parameter.

N: The number of individuals.

The print.summary.fits_ids_dm function displays the summary object in a formatted manner.

Examples

# get an auxiliary object of type fits_ids_dm for demonstration purpose
all_fits <- get_example_fits_ids()
sum_obj <- summary(all_fits)
print(sum_obj, round_digits = 2)
#> Fit Procedure Name: aux_example
#> N Individuals: 3 
#> 
#> Parameter Summary: comp 
#>          muc    b non_dec sd_non_dec  tau a    A alpha peak_l
#> mean    5.30 0.48    0.32       0.03 0.07 2 0.13  4.57   0.07
#> std_err 0.32 0.06    0.01       0.01 0.02 0 0.03  1.23   0.02
#> 
#> Parameter Summary: incomp 
#>          muc    b non_dec sd_non_dec  tau a     A alpha peak_l
#> mean    5.30 0.48    0.32       0.03 0.07 2 -0.13  4.57   0.07
#> std_err 0.32 0.06    0.01       0.01 0.02 0  0.03  1.23   0.02
#> 
#> 
#> Parameter Space:
#>       muc   b non_dec sd_non_dec  tau    A alpha
#> lower   1 0.2     0.1      0.005 0.02 0.02     3
#> upper   7 1.0     0.6      0.100 0.30 0.30     8
#> 
#> -------
#> Fitted Model Type: dmc_dm, drift_dm
#> Time of (Last) Call: 2025-February-21_09-22